updates for python3
