+++ .travis.yml
@@ -2,8 +2,9 @@
 language:
   - python
 
-# specify python versions
+# specify python version(s)
 python:
+  - "3.4"
   - "2.7"
 
 # install dependencies
@@ -15,11 +16,11 @@ script:
   - nosetests
 
 # deploy
-deploy:
-  provider: heroku
-  api_key:
-    secure: XEtRjqqP6ZgiZKDi/OjSyMhJIPVPUzmX+5899LXfm2eyMtItYqFouCtBAOmBU5bWnZiwgXdB9l9ra2yiC+321TET1WVJe1WHDTFW1rzIpWS7t/ZIJYPhmwcYxa02kzLQZsrDZH/BW/cUuoHlOhmGSpuqO6SO1h1XnzwqHVe68vw=
-  app: flasktaskr
-  on:
-    branch: master
-    repo: realpython/flasktaskr_project
+# deploy:
+#   provider: heroku
+#   api_key:
+#     secure: XEtRjqqP6ZgiZKDi/OjSyMhJIPVPUzmX+5899LXfm2eyMtItYqFouCtBAOmBU5bWnZiwgXdB9l9ra2yiC+321TET1WVJe1WHDTFW1rzIpWS7t/ZIJYPhmwcYxa02kzLQZsrDZH/BW/cUuoHlOhmGSpuqO6SO1h1XnzwqHVe68vw=
+#   app: flasktaskr
+#   on:
+#     branch: master
+#     repo: realpython/flasktaskr_project
