+++ project/_config.py
@@ -1,9 +1,10 @@
-# config.py
+# project/_config.py
 
 
 import os
 
-# grabs the folder where the script runs
+
+# grab the folder where this script lives
 basedir = os.path.abspath(os.path.dirname(__file__))
 
 DATABASE = 'flasktaskr.db'
@@ -11,7 +12,7 @@ CSRF_ENABLED = True
 SECRET_KEY = 'my_precious'
 DEBUG = True
 
-# defines the full path for the database
+# define the full path for the database
 DATABASE_PATH = os.path.join(basedir, DATABASE)
 
 # the database uri
