+++ project/api/__init__.py
@@ -1,0 +1,1 @@
+# project/api/__init__.py
