+++ project/static/css/main.css
@@ -1,5 +1,5 @@
 body {
-  padding: 60px 0px;
+  padding: 60px 0;
   background: #ffffff;
 }
