+++ project/tasks/forms.py
@@ -1,15 +1,15 @@
-# /project/tasks/forms.py
+# project/users/forms.py
 
 
 from flask_wtf import Form
-from wtforms import TextField, DateField, IntegerField, \
+from wtforms import StringField, DateField, IntegerField, \
     SelectField
 from wtforms.validators import DataRequired
 
 
 class AddTaskForm(Form):
     task_id = IntegerField('Priority')
-    name = TextField('Task Name', validators=[DataRequired()])
+    name = StringField('Task Name', validators=[DataRequired()])
     due_date = DateField(
         'Date Due (mm/dd/yyyy)',
         validators=[DataRequired()], format='%m/%d/%Y'
