--- project/tasks/templates/tasks.html
@@ -1,122 +1,0 @@
-<!-- /project/tasks/templates/tasks.html -->
-
-
-{% extends "template.html" %}
-
-{% block content %}
-
-  <h1>Welcome to FlaskTaskr</h1>
-  <div class="add-task">
-    <div class="lead">Add a new task:</div>
-      <form action="{{ url_for('tasks.new_task') }}" method="post">
-        {{ form.csrf_token }}
-        <p>
-        {{ form.name(placeholder="name") }}
-        <span class="error">
-          {% if form.name.errors %}
-            {% for error in form.name.errors %}
-              {{ error }}
-            {% endfor %}
-          {% endif %}
-        </span>
-        </p>
-        <p>
-        {{ form.due_date(placeholder="due date") }}
-        <span class="error">
-          {% if form.due_date.errors %}
-            {% for error in form.due_date.errors %}
-              {{ error }}
-            {% endfor %}
-          {% endif %}
-        </span>
-        </p>
-        <p>
-        {{ form.priority.label }}
-        {{ form.priority(class="form-control",style="max-width:60px;") }}
-        <span class="error">
-          {% if form.priority.errors %}
-            {% for error in form.priority.errors %}
-              {{ error }}
-            {% endfor %}
-          {% endif %}
-        </span>
-        </p>
-        <br>
-        <p><button class="btn btn-sm btn-success" type="submit">Submit</button></p>
-      </form>
-  </div>
-<div class="entries">
-  <br>
-  <br>
-  <div class="lead">Open tasks</div>
-  <br>
-  <table class="table table-striped">
-    <thead>
-      <tr>
-        <th>Task Name</th>
-        <th>Due Date</th>
-        <th>Posted Date</th>
-        <th>Priority</th>
-        <th>Posted By</th>
-        <th>Actions</th>
-      </tr>
-    </thead>
-    <tbody>
-    {% for o in open_tasks %}
-      <tr>
-        <td>{{ o.name }}</td>
-        <td>{{ o.due_date }}</td>
-        <td>{{ o.posted_date }}</td>
-        <td>{{ o.priority }}</td>
-        <td>{{ o.poster.name }}</td>
-        <td>
-          {% if o.poster.name == session.name or session.role == "admin" %}
-            <a href="{{ url_for('tasks.delete_entry', task_id = o.task_id) }}">Delete</a>  -
-            <a href="{{ url_for('tasks.complete', task_id = o.task_id) }}">Mark as Complete</a>
-          {% else %}
-            <span>N/A</span>
-          {% endif %}
-        </td>
-      </tr>
-    {% endfor %}
-    </tbody>
-  </table>
-</div>
-<br>
-<br>
-<div class="entries">
-  <div class="lead">Closed tasks</div>
-  <br>
-  <table class="table table-striped">
-    <thead>
-      <tr>
-        <th>Task Name</th>
-        <th>Due Date</th>
-        <th>Posted Date</th>
-        <th>Priority</th>
-        <th>Posted By</th>
-        <th>Actions</th>
-      </tr>
-    </thead>
-    <tbody>
-    {% for c in closed_tasks %}
-      <tr>
-        <td>{{ c.name }}</td>
-        <td>{{ c.due_date }}</td>
-        <td>{{ c.posted_date }}</td>
-        <td>{{ c.priority }}</td>
-        <td>{{ c.poster.name }}</td>
-        <td>
-          {% if c.poster.name == session.name or session.role == "admin" %}
-           <a href="{{ url_for('tasks.delete_entry', task_id = c.task_id) }}">Delete</a>
-          {% else %}
-            <span>N/A</span>
-          {% endif %}
-        </td>
-      </tr>
-    {% endfor %}
-    <tbody>
-  </table>
-</div>
-
-{% endblock %}
