+++ project/templates/500.html
@@ -1,6 +1,4 @@
-<!-- project/templates/500.html -->
-
-{% extends "template.html" %}
+{% extends "_base.html" %}
 
 {% block content %}
