+++ project/templates/_base.html
@@ -1,6 +1,3 @@
-<!-- project/templates/template.html -->
-
-
 <!DOCTYPE html>
 <html lang="en">
   <head>
@@ -11,10 +8,11 @@
     <meta name="author" content="">
     <title>Welcome to FlaskTaskr!!!</title>
     <!-- styles -->
-    <link href="http://maxcdn.bootstrapcdn.com/bootswatch/3.2.0/yeti/bootstrap.min.css" rel="stylesheet">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
+    <link href="//maxcdn.bootstrapcdn.com/bootswatch/3.3.4/yeti/bootstrap.min.css" rel="stylesheet">
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/main.css') }}">
   </head>
   <body>
+
     <div class="navbar navbar-inverse navbar-fixed-top" role="navigation">
       <div class="container">
         <div class="navbar-header">
@@ -45,9 +43,11 @@
 
     <div class="container">
       <div class="content">
+
         {% for message in get_flashed_messages() %}
-          <div class="flash">{{ message }}</div>
+          <div class="alert alert-success">{{ message }}</div>
         {% endfor %}
+
         {% if error %}
           <div class="error"><strong>Error:</strong> {{ error }}</div>
         {% endif %}
@@ -58,11 +58,12 @@
       </div>
       <div class="footer">
         <hr>
-        <p>&copy; <a href="http://www.realpython.com">Real Python</a></p>
+        <p>&copy; <a href="https://www.realpython.com">Real Python</a></p>
       </div>
     </div><!-- /.container -->
+
     <!-- scripts -->
-    <script src="http://code.jquery.com/jquery-1.10.2.min.js"></script>
-    <script src="http://netdna.bootstrapcdn.com/bootstrap/3.2.0/js/bootstrap.min.js"></script>
+    <script src="//code.jquery.com/ui/1.11.4/jquery-ui.min.js"></script>
+    <script src="//maxcdn.bootstrapcdn.com/bootstrap/3.3.4/js/bootstrap.min.js"></script>
   </body>
 </html>
