+++ project/templates/register.html
@@ -1,57 +1,53 @@
-<!-- /project/users/templates/register.html -->
-
-
-{% extends "template.html" %}
-
+{% extends "_base.html" %}
 {% block content %}
 
 <h1>Welcome to FlaskTaskr.</h1>
-<form class="form-signin" role="form" method="post" action="">
-  <div class="lead">Please register to start a task list</div>
+<h3>Please register to access the task list.</h3>
+<form method="POST" action="/register/">
   {{ form.csrf_token }}
   <p>
     {{ form.name(placeholder="name") }}
-    <span class="error">
-      {% if form.name.errors %}
+    {% if form.name.errors %}
+      <span class="error">
         {% for error in form.name.errors %}
           {{ error }}
         {% endfor %}
-      {% endif %}
-    </span>
+      </span>
+    {% endif %}
   </p>
   <p>
     {{ form.email(placeholder="email address") }}
-    <span class="error">
-      {% if form.email.errors %}
+    {% if form.name.errors %}
+      <span class="error">
         {% for error in form.email.errors %}
           {{ error }}
         {% endfor %}
-      {% endif %}
-    </span>
+      </span>
+    {% endif %}
   </p>
   <p>
     {{ form.password(placeholder="password") }}
-    <span class="error">
-      {% if form.password.errors %}
+    {% if form.name.errors %}
+      <span class="error">
         {% for error in form.password.errors %}
           {{ error }}
         {% endfor %}
-      {% endif %}
-    </span>
+      </span>
+    {% endif %}
   </p>
+  <p>
     {{ form.confirm(placeholder="confirm password") }}
-    <span class="error">
-      {% if form.confirm.errors %}
+    {% if form.name.errors %}
+      <span class="error">
         {% for error in form.confirm.errors %}
           {{ error }}
         {% endfor %}
-      {% endif %}
-    </span>
+      </span>
+    {% endif %}
   </p>
   <button class="btn btn-sm btn-success" type="submit">Register</button>
   <br>
-  <br>
-  <p><em>Already registered?</em> Click <a href="{{ url_for('users.login') }}">here</a> to login.</p>
+  <p><em>Already registered?</em> Click <a href="/">here</a> to login.</p>
 </form>
 
 {% endblock %}
