+++ project/users/views.py
@@ -1,16 +1,17 @@
-# /project/users/views.py
+# project/users/views.py
 
 
 #################
 #### imports ####
 #################
 
-from flask import flash, redirect, render_template, request, \
-    session, url_for, Blueprint
+from functools import wraps
+from flask import flash, redirect, render_template, \
+    request, session, url_for, Blueprint
 from sqlalchemy.exc import IntegrityError
-from forms import RegisterForm, LoginForm
+
+from .forms import RegisterForm, LoginForm
 from project import db, bcrypt
-from project.views import login_required
 from project.models import User
 
 
@@ -18,12 +19,23 @@ from project.models import User
 #### config ####
 ################
 
-users_blueprint = Blueprint(
-    'users', __name__,
-    url_prefix='/users',
-    template_folder='templates',
-    static_folder='static'
-)
+users_blueprint = Blueprint('users', __name__)
+
+
+##########################
+#### helper functions ####
+##########################
+
+
+def login_required(test):
+    @wraps(test)
+    def wrap(*args, **kwargs):
+        if 'logged_in' in session:
+            return test(*args, **kwargs)
+        else:
+            flash('You need to login first.')
+            return redirect(url_for('users.login'))
+    return wrap
 
 
 ################
@@ -37,7 +49,7 @@ def logout():
     session.pop('user_id', None)
     session.pop('role', None)
     session.pop('name', None)
-    flash('You are logged out.')
+    flash('Goodbye!')
     return redirect(url_for('users.login'))
 
 
@@ -48,30 +60,17 @@ def login():
     if request.method == 'POST':
         if form.validate_on_submit():
             user = User.query.filter_by(name=request.form['name']).first()
-            if user is None:
-                error = 'Invalid username or password.'
-                return render_template(
-                    "login.html",
-                    form=form,
-                    error=error
-                )
-            elif bcrypt.check_password_hash(
-                user.password, request.form['password']
-            ):
+            if user is not None and bcrypt.check_password_hash(
+                    user.password, request.form['password']):
                 session['logged_in'] = True
                 session['user_id'] = user.id
                 session['role'] = user.role
                 session['name'] = user.name
-                flash('You are logged in. Go Crazy.')
+                flash('Welcome!')
                 return redirect(url_for('tasks.tasks'))
-        else:
-            return render_template(
-                "login.html",
-                form=form,
-                error=error
-            )
-    if request.method == 'GET':
-        return render_template('login.html', form=form)
+            else:
+                error = 'Invalid username or password.'
+    return render_template('login.html', form=form, error=error)
 
 
 @users_blueprint.route('/register/', methods=['GET', 'POST'])
@@ -91,9 +90,6 @@ def register():
                 flash('Thanks for registering. Please login.')
                 return redirect(url_for('users.login'))
             except IntegrityError:
-                error = 'Sorry that username and/or email error already exist.'
+                error = 'That username and/or email already exist.'
                 return render_template('register.html', form=form, error=error)
-        else:
-            return render_template('register.html', form=form, error=error)
-    if request.method == 'GET':
-        return render_template('register.html', form=form)
+    return render_template('register.html', form=form, error=error)
