+++ requirements.txt
@@ -1,15 +1,18 @@
+Fabric==1.10.1
 Flask==0.10.1
-Flask-Bcrypt==0.6.0
-Flask-SQLAlchemy==1.0
-Flask-WTF==0.9.4
+Flask-Bcrypt==0.6.2
+Flask-SQLAlchemy==2.0
+Flask-WTF==0.11
 Jinja2==2.7.3
 MarkupSafe==0.23
-SQLAlchemy==0.9.7
-WTForms==1.0.5
-Werkzeug==0.9.6
+SQLAlchemy==0.9.9
+WTForms==2.0.2
+Werkzeug==0.10.1
 coverage==3.7.1
-gunicorn==19.1.0
+ecdsa==0.13
+gunicorn==19.3.0
 itsdangerous==0.24
-nose==1.3.3
-py-bcrypt==0.4
-wsgiref==0.1.2
+nose==1.3.4
+paramiko==1.15.2
+pycrypto==2.6.1
+python-bcrypt==0.3.1
